Disable prefix detection
This is an attempt to fix #41.

The idea that this line messes things up:
https://github.com/nushell/nushell/blob/eb2b46992271588d5d59bd34c60800862cfcd2de/crates/nu-protocol/src/eval_const.rs#L314

Further details can be found here:
https://rattler.build/latest/debugging_builds/#understanding-relocatability

I am not 100% if that fixes the issue, since it isn't perfectly
reproducible for me. Without this change 1/3 builds have this problem,
with this change I haven't encountered it.

Without Prefix replacement we can't use auto-loaded scripts in the
future, but it's still better than this bug. We should re-evaluate in
the future, assuming this is actually the problem